Hardcore Mode, Spectating bug fixes
Added hardcore mode which allows players to remove the gameover state if they choose for a more casual experience.

Made scoreboard reset on hardcore fail.

If hardcore and respawning are both disabled it will force respawning on upon a death to avoid a soft-lock.

Fixed errors from spectating in singleplayer.